requirements.txt
flask
python-dotenv
pymupdf
tiktoken
openai
pinecone-client
